+++ test/Gaa.Extensions.Mediator.Test/Features/ExampleRequest.cs
@@ -1,0 +1,13 @@
+namespace Gaa.Extensions.Test.Features;
+
+/// <summary>
+/// Пример запроса.
+/// </summary>
+public class ExampleRequest
+    : IRequest
+{
+    /// <summary>
+    /// Текст с сообщением.
+    /// </summary>
+    public string Message { get; init; } = "Test message from request!";
+}