+++ test/Gaa.Extensions.Mediator.Test/Features/ExampleRequestWithResponse.cs
@@ -1,0 +1,13 @@
+namespace Gaa.Extensions.Test.Features;
+
+/// <summary>
+/// Пример запроса.
+/// </summary>
+public class ExampleRequestWithResponse
+    : IRequest<ExampleResponse>
+{
+    /// <summary>
+    /// Текст с сообщением.
+    /// </summary>
+    public string Message { get; init; } = "Test message from response!";
+}