+++ test/Gaa.Extensions.Mediator.Test/Features/ExampleResponse.cs
@@ -1,0 +1,12 @@
+namespace Gaa.Extensions.Test.Features;
+
+/// <summary>
+/// Пример ответа.
+/// </summary>
+public class ExampleResponse
+{
+    /// <summary>
+    /// Текст с сообщением.
+    /// </summary>
+    public string Message { get; init; } = "Test message from response!";
+}